Moving all preferences-related stores to preferences dir
