+++ src/routes/preferences/misc/+page.svelte
@@ -4,7 +4,7 @@
     import FormContainer from "$components/ui/forms/FormContainer.svelte";
     import FormControl from "$components/ui/forms/FormControl.svelte";
     import CheckboxField from "$components/ui/forms/CheckboxField.svelte";
-    import { fullScreenViewerEnabled } from "$stores/misc-preferences";
+    import { fullScreenViewerEnabled } from "$stores/preferences/misc";
 </script>
 
 <Menu>
