+++ src/stores/preferences/misc.ts
