+++ src/stores/preferences/search.ts
